shared array
I made some functions that can mess with the shared array. I began to use them with the Amplifier and the Launcher. It is scalable through a variable that says how many indexes to use. The bits are stored with the exact locations and the rest are what to communicate. Either the location of newfound objects or dest_launcher will tell all launchers to go somewhere. These methods don't have to be used, I am familiarizing myself with these shared arrays.